+++ src/main/java/hhplus/booking/app/point/interfaces/api/PointController.java
@@ -1,9 +1,14 @@
-package hhplus.booking.interfaces.point;
+package hhplus.booking.app.point.interfaces.api;
 
-import hhplus.booking.interfaces.point.dto.PointInfo;
+import hhplus.booking.app.point.interfaces.api.dto.PointInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
