+++ src/main/java/hhplus/booking/app/user/interfaces/api/dto/UserQueueInfo.java
@@ -1,0 +1,6 @@
+package hhplus.booking.app.user.interfaces.api.dto;
+
+public class UserQueueInfo {
+    public record Input(String authorizationHeader, Long userId) {}
+    public record Output(UserTokenInfo userTokenInfo, Long queueRank) {}
+}
