+++ src/main/java/hhplus/booking/app/user/interfaces/api/dto/UserTokenInfo.java
@@ -1,0 +1,6 @@
+package hhplus.booking.app.user.interfaces.api.dto;
+
+public record UserTokenInfo (
+        String tokenValue,
+        Long userId
+){ }
