--- src/main/java/hhplus/booking/interfaces/user/dto/UserQueueInfo.java
@@ -1,6 +1,0 @@
-package hhplus.booking.interfaces.user.dto;
-
-public class UserQueueInfo {
-    public record Input() {}
-    public record Output(Long userId, Long queueRank) {}
-}
