--- src/main/java/hhplus/booking/interfaces/user/dto/UserTokenInfo.java
@@ -1,6 +1,0 @@
-package hhplus.booking.interfaces.user.dto;
-
-public class UserTokenInfo {
-    public record Input(Long userId) {}
-    public record Output(String tokenValue) {}
-}
